// src/db/schema.d.ts
export type Generated<T> =
  T extends ColumnType<infer S, infer I, infer U>
    ? ColumnType<S, I | undefined, U>
    : ColumnType<T, T | undefined, T>;

export type Interval = ColumnType<
  IPostgresInterval,
  IPostgresInterval | number | string,
  IPostgresInterval | number | string
>;

export type Json = JsonValue;

export type JsonArray = JsonValue[];

export type JsonObject = {
  [x: string]: JsonValue | undefined;
};

export type JsonPrimitive = boolean | number | string | null;

export type JsonValue = JsonArray | JsonObject | JsonPrimitive;

export type Numeric = ColumnType<string, number | string, number | string>;

export type Point = {
  x: number;
  y: number;
};

export type Timestamp = ColumnType<Date, Date | string, Date | string>;

export interface Airplanes {
  /**
   * Airplane code, IATA
   */
  airplane_code: string | null;
  /**
   * Airplane model
   */
  model: string | null;
  /**
   * Maximum flight range, km
   */
  range: number | null;
  /**
   * Cruise speed, km/h
   */
  speed: number | null;
}

export interface AirplanesData {
  /**
   * Airplane code, IATA
   */
  airplane_code: string;
  /**
   * Airplane model
   */
  model: Json;
  /**
   * Maximum flight range, km
   */
  range: number;
  /**
   * Cruise speed, km/h
   */
  speed: number;
}

export interface Airports {
  /**
   * Airport code, IATA
   */
  airport_code: string | null;
  /**
   * Airport name
   */
  airport_name: string | null;
  /**
   * City
   */
  city: string | null;
  /**
   * Airport coordinates (longitude and latitude)
   */
  coordinates: Point | null;
  /**
   * Country
   */
  country: string | null;
  /**
   * Airport time zone
   */
  timezone: string | null;
}

export interface AirportsData {
  /**
   * Airport code, IATA
   */
  airport_code: string;
  /**
   * Airport name
   */
  airport_name: Json;
  /**
   * City
   */
  city: Json;
  /**
   * Airport coordinates (longitude and latitude)
   */
  coordinates: Point;
  /**
   * Country
   */
  country: Json;
  /**
   * Airport time zone
   */
  timezone: string;
}

export interface BoardingPasses {
  /**
   * Boarding pass number
   */
  boarding_no: number | null;
  /**
   * Boarding time
   */
  boarding_time: Timestamp | null;
  /**
   * Flight ID
   */
  flight_id: number;
  /**
   * Seat number
   */
  seat_no: string;
  /**
   * Ticket number
   */
  ticket_no: string;
}

export interface Bookings {
  /**
   * Booking date
   */
  book_date: Timestamp;
  /**
   * Booking number
   */
  book_ref: string;
  /**
   * Total booking amount
   */
  total_amount: Numeric;
}

export interface Flights {
  /**
   * Actual arrival time
   */
  actual_arrival: Timestamp | null;
  /**
   * Actual departure time
   */
  actual_departure: Timestamp | null;
  /**
   * Flight ID
   */
  flight_id: Generated<number>;
  /**
   * Route number
   */
  route_no: string;
  /**
   * Scheduled arrival time
   */
  scheduled_arrival: Timestamp;
  /**
   * Scheduled departure time
   */
  scheduled_departure: Timestamp;
  /**
   * Flight status
   */
  status: string;
}

export interface Routes {
  /**
   * Airplane code, IATA
   */
  airplane_code: string;
  /**
   * Airport of arrival
   */
  arrival_airport: string;
  /**
   * Days of week array
   */
  days_of_week: number[];
  /**
   * Airport of departure
   */
  departure_airport: string;
  /**
   * Estimated duration
   */
  duration: Interval;
  /**
   * Route number
   */
  route_no: string;
  /**
   * Scheduled local time of departure
   */
  scheduled_time: string;
  /**
   * Period of validity
   */
  validity: string;
}

export interface Seats {
  /**
   * Airplane code, IATA
   */
  airplane_code: string;
  /**
   * Travel class
   */
  fare_conditions: string;
  /**
   * Seat number
   */
  seat_no: string;
}

export interface Segments {
  /**
   * Travel class
   */
  fare_conditions: string;
  /**
   * Flight ID
   */
  flight_id: number;
  /**
   * Travel price
   */
  price: Numeric;
  /**
   * Ticket number
   */
  ticket_no: string;
}

export interface Tickets {
  /**
   * Booking number
   */
  book_ref: string;
  /**
   * Outbound flight?
   */
  outbound: boolean;
  /**
   * Passenger ID
   */
  passenger_id: string;
  /**
   * Passenger name
   */
  passenger_name: string;
  /**
   * Ticket number
   */
  ticket_no: string;
}

export interface Timetable {
  /**
   * Actual arrival time
   */
  actual_arrival: Timestamp | null;
  /**
   * Actual arrival time in airport's timezone
   */
  actual_arrival_local: Timestamp | null;
  /**
   * Actual departure time
   */
  actual_departure: Timestamp | null;
  /**
   * Actual departure time in airport's timezone
   */
  actual_departure_local: Timestamp | null;
  /**
   * Airplane code, IATA
   */
  airplane_code: string | null;
  /**
   * Airport of arrival
   */
  arrival_airport: string | null;
  /**
   * Airport of departure
   */
  departure_airport: string | null;
  /**
   * Flight ID
   */
  flight_id: number | null;
  /**
   * Route number
   */
  route_no: string | null;
  /**
   * Scheduled arrival time
   */
  scheduled_arrival: Timestamp | null;
  /**
   * Scheduled arrival time in airport's timezone
   */
  scheduled_arrival_local: Timestamp | null;
  /**
   * Scheduled departure time
   */
  scheduled_departure: Timestamp | null;
  /**
   * Scheduled departure time in airport's timezone
   */
  scheduled_departure_local: Timestamp | null;
  /**
   * Flight status
   */
  status: string | null;
}

export interface DB {
  airplanes: Airplanes;
  airplanes_data: AirplanesData;
  airports: Airports;
  airports_data: AirportsData;
  boarding_passes: BoardingPasses;
  bookings: Bookings;
  flights: Flights;
  routes: Routes;
  seats: Seats;
  segments: Segments;
  tickets: Tickets;
  timetable: Timetable;
}